+++ packages/medusa_api_client/lib/src/models/get_draft_orders_query_parameters.dart
@@ -161,11 +161,11 @@ abstract class GetDraftOrdersQueryParameters
 
     /// offset
     @JsonKey(name: GetDraftOrdersQueryParameters.offsetKey)
-    required double offset,
+    required int offset,
 
     /// limit
     @JsonKey(name: GetDraftOrdersQueryParameters.limitKey)
-    required double limit,
+    required int limit,
 
     /// order
     @JsonKey(name: GetDraftOrdersQueryParameters.orderKey)
