+++ packages/medusa_api_client/lib/src/models/get_orders_id_line_items_query_parameters.dart
@@ -131,11 +131,11 @@ abstract class GetOrdersIdLineItemsQueryParameters
 
     /// limit
     @JsonKey(name: GetOrdersIdLineItemsQueryParameters.limitKey)
-    required double limit,
+    required int limit,
 
     /// offset
     @JsonKey(name: GetOrdersIdLineItemsQueryParameters.offsetKey)
-    required double offset,
+    required int offset,
 
     /// order
     @JsonKey(name: GetOrdersIdLineItemsQueryParameters.orderKey)
