+++ packages/medusa_api_client/lib/src/models/get_price_lists_query_parameters.dart
@@ -111,10 +111,10 @@ abstract class GetPriceListsQueryParameters
 
     /// offset
     @JsonKey(name: GetPriceListsQueryParameters.offsetKey)
-    required double offset,
+    required int offset,
 
     /// limit
-    @JsonKey(name: GetPriceListsQueryParameters.limitKey) required double limit,
+    @JsonKey(name: GetPriceListsQueryParameters.limitKey) required int limit,
 
     /// order
     @JsonKey(name: GetPriceListsQueryParameters.orderKey) required String order,
