+++ packages/medusa_api_client/lib/src/models/get_products_id_variants_query_parameters.dart
@@ -104,11 +104,11 @@ abstract class GetProductsIdVariantsQueryParameters
 
     /// offset
     @JsonKey(name: GetProductsIdVariantsQueryParameters.offsetKey)
-    required double offset,
+    required int offset,
 
     /// limit
     @JsonKey(name: GetProductsIdVariantsQueryParameters.limitKey)
-    required double limit,
+    required int limit,
 
     /// order
     @JsonKey(name: GetProductsIdVariantsQueryParameters.orderKey)
